Feat: [BADA-169] StroeDevice 관련 api 및 model 자리 원위치 #96
