--- src/entities/devices/lib/types.ts
@@ -1,6 +1,0 @@
-export interface StoreDevice {
-  storeDeviceId: number;
-  deviceName: string;
-  dataCapacity: number;
-  imageUrl: string;
-}
