+++ src/pages/rental/map/lib/types.ts
@@ -4,5 +4,9 @@ export interface Store {
   latitude: number;
   longititude: number;
 }
-
-
+export interface StoreDevice {
+  storeDeviceId: number;
+  deviceName: string;
+  dataCapacity: number;
+  imageUrl: string;
+}
